FEAT

+++ hackweek-backend/dtos/RatingCriterionDTO.cs
@@ -2,7 +2,7 @@
 {
     public class RatingCriterionDTO
     {
-        public int Grade { set; get; }
+        public double Grade { set; get; }
         public int CriterionId { get; set; }
     }
 }
